Fixed and enhanced OAuth2 get token workflow

+++ oauth2/token.go
@@ -4,6 +4,6 @@ type Token struct {
 	IdToken     string `json:"id_token"`
 	AccessToken string `json:"access_token"`
 	TokenType   string `json:"token_type"`
-	ExpiresIn   uint32 `json:"expires_in"`
+	ExpiresIn   uint64 `json:"expires_in"`
 	Scope       string `json:"scope"`
 }
